+++ src/services/homebridge/scheduleService.ts
@@ -1,8 +1,8 @@
 import { API, CharacteristicSetCallback, PlatformAccessory } from 'homebridge';
-import { AutomowerContext } from '../automowerAccessory';
+import { AutomowerContext } from '../../automowerAccessory';
 
-import { MowerControlService } from './automower/mowerControlService';
-import { SwitchService } from './switchService';
+import { MowerControlService } from '../automower/mowerControlService';
+import { AbstractSwitchService } from './abstractSwitchService';
 
 /**
  * A service which manages the schedule enablement of an automower.
@@ -14,7 +14,7 @@ export interface ScheduleService {
     init(): void;
 }
 
-export class ScheduleServiceImpl extends SwitchService implements ScheduleService {    
+export class ScheduleServiceImpl extends AbstractSwitchService implements ScheduleService {    
     public constructor(private controlService: MowerControlService, accessory: PlatformAccessory<AutomowerContext>, api: API) {
         super('Schedule', accessory, api);
     }
